move shape to shape.h
