+++ CMakeLists.txt
@@ -11,6 +11,7 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${bahnhof_SOURCE_DIR}/cmake")
 add_executable(bahnhof
     src/main.cpp
     src/math.cpp
+    src/shape.cpp
     src/graphics/rendering.cpp
     src/graphics/graphics.cpp
     src/graphics/sprite.cpp
